refactor: move 'reset internal wagmi state' to the useAppChainManager, use the useChainId
